Properly stub out snail::window during headless testing
