+++ src/snail/hsp/headless.cpp
@@ -1,3 +1,4 @@
+#include "../application.hpp"
 #include <iostream>
 
 namespace elona
@@ -101,8 +102,9 @@ void line(int, int, int, int, const snail::color&)
 {
 }
 
-void title(const std::string&, const std::string&, window::fullscreen_mode_t)
+void title(const std::string& title_str, const std::string&, window::fullscreen_mode_t)
 {
+    application::instance().initialize(title_str);
 }
